Module mandalas::util

Various utility functions.

Statics

FLAG_ACE

Functions

cmyk_to_rgb
Translate a CMYK colour point into an RGB colour point.

distance
Calculate the distance between two specified 3-dimensional points.

flag_hard_alpha_to_rgb
Translate a colour point and alpha into an RGB colour point from the specified flag colourset.

hsl_to_rgb
Translate an HSL colour point into an RGB colour point.

prgb1555_to_rgb
Translate a PRGB1555 colour point into an RGB colour point.

prgb1555_least6_to_rgb
Translate a PRGB1555 colour point into an RGB colour point.

proygbiv_to_rgb
Translate a PROYGBIV colour point into an RGB colour point according to sampling the svg on wikipedia.

rgb_to_sepia
Translate an RGB colour point into a Sepia tone.

roygbp_to_rgb
Translate a ROYGBP colour point into an RGB colour point according to sampling the svg on wikipedia.

rppdiblctgllyaodbgb_to_rgb
Translate a RPPDIBLCTGLLYAODBGB colour point into an RGB colour point according to this PNG.

separated_number
Convert the specified number to a string in the form "1'234'567"

yuv_to_rgb
Translate a YUV colour point (translated to 0–1 as opposed to -0.5–0.5) into an RGB colour point according to the BT.709 recommendation.